requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
httpx>=0.27.0
python-dotenv>=1.0.1
hubspot-api-client>=11.1.0
